requirements.txt
asgiref==3.3.4
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.3
click==8.0.1
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
legacy==0.1.6
MarkupSafe==2.0.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
Werkzeug==2.0.1
whitenoise==5.2.0
